clean and add useUpdatePLugin hook
